[Person] refines and optimizes db schema and related data objects

+++ person/src/main/java/naatsms/person/dto/AddressDto.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 public record AddressDto(
         UUID id,
-        Long countryId,
+        Integer countryId,
         String address,
         String zipCode,
         String city,
